Shorten middleware module name
