+++ lib/permit_absinthe/middleware/load_and_authorize.ex
@@ -1,67 +1,19 @@
 defmodule Permit.Absinthe.Middleware.LoadAndAuthorize do
   @moduledoc """
-  Middleware for loading and authorizing resources in Absinthe. Uses the raw
-  resolver function defined in `Permit.Absinthe.Resolvers.LoadAndAuthorize` to
-  put the resolution outcome in `:loaded_resource` or `:loaded_resources`
-  (depending on whether it's an index-like or a single-item action).
-
-  Useful in mutations where it's not enough to just load a resource, or whenever
-  you otherwise need to process it after loading.
-
-  ## Usage, mechanism and configuration
-
-  See `Permit.Absinthe.Resolvers.LoadAndAuthorize` for details on resolution of
-  authorized records in Permit.Absinthe. The middleware delegates
-
-  ### Example
-
-  ```elixir
-  mutation do
-    @desc "Update an article"
-    field :update_article, :article do
-      permit action: :update
-
-      arg(:id, non_null(:id))
-      arg(:name, non_null(:string))
-      arg(:content, non_null(:string))
-
-      middleware Permit.Absinthe.Middleware.LoadAndAuthorize
-
-      resolve(fn _, %{name: name, content: content}, %{context: context} ->
-        case Blog.Content.update_article(context.loaded_resource, %{name: name, content: content}) do
-          {:ok, article} ->
-            {:ok, article}
-
-          {:error, changeset} ->
-            {:error, "Could not update article"}
-        end
-      end)
-    end
-  end
-  ```
+  Deprecated. Use `Permit.Absinthe.Middleware` instead.
   """
-  alias Permit.Absinthe.Resolvers.LoadAndAuthorize
+
+  @deprecated "Use Permit.Absinthe.Middleware instead"
 
   @behaviour Absinthe.Middleware
 
   @impl true
-  def call(resolution, _) do
-    case LoadAndAuthorize.load_and_authorize(
-           resolution.arguments,
-           resolution
-         ) do
-      {:ok, resource} ->
-        key =
-          case resource do
-            list when is_list(list) -> :loaded_resources
-            _ -> :loaded_resource
-          end
+  def call(resolution, opts) do
+    IO.warn(
+      "Permit.Absinthe.Middleware.LoadAndAuthorize is deprecated, use Permit.Absinthe.Middleware instead",
+      []
+    )
 
-        new_context = Map.put(resolution.context, key, resource)
-        %{resolution | context: new_context}
-
-      {:error, error} ->
-        Absinthe.Resolution.put_result(resolution, {:error, error})
-    end
+    Permit.Absinthe.Middleware.call(resolution, opts)
   end
 end
